Pin berkshelf to a version prior to requiring net-ssh 3.0.2
Supermarket rails app requires the chef gem for version number logic. That
chef gem is Gemfile.lock'd to 12.0 which requires a 2.x version of net-ssh.
Riding berkshelf master in this software project meant the package build
pulled in a newer net-ssh in addition to the 2.9.2 that the rails app
needed. And then things got upset. Can't upgrade the chef gem immediately
because newer chef requires newer rspec which breaks our tests.

Good times.