Use rb-readline gem to fix the Rails console
We upgraded omnibus-software to a version whose Ruby no longer had readline
compiled within. This adds the rb-readline gem to provide the same functionality
with only Ruby and UNIX commands.

Needed to update the PATH in the -ctl console command so that utilities like
stty required by rb-readline would be available.