Add sensible default for supermarket/fieri to talk to itself
localhost:13000 is the default unicorn listening socket. This default
assumes that the Sidekiq workers are running on a host that is also running
the Rails web app which is—as of this commit—the only way to run those
services. In a world where workers can be run on a host separate from the
web app, these attributes will need to be overridden, probably with the URI
used by end-users.