Fix Quality tab not displaying successful foodcritic feedback.
foodcritic_failure will be nil if fieri hasn't posted back results. If
fieri has posted results, foodcritic_failure will be true or false. false
means success (argh, double negatives!), but false is also not `present?`
and so the test for presence fails and the view thinks there are no
results.

This flips the test from `present?` to `nil?` and flips the if/else
clauses. This was the smallest change to fix the display of foodcritic
success.