Fix index name too long in an old migration.
Setting a specific name for the index keeps it from exceeding
PG's 63 character limit. When the table is renamed later, the
index is also renamed so that it follows Rails convention and
subsequent modifications to the table will update the index.

schema.rb updated by Rails to use the new force: :cascade
default in SchemaDumper[1].

> SchemaDumper uses force: :cascade on create_table. This
> makes it possible to reload a schema when foreign keys are
> in place.

schema.rb also was updated by Rails 4.2 migrations to create
new indexes not previously generated by the framework. These
indexes are on columns have a uniqueness constraint, are
used as a parameter in a scope, or are a reference to an id
in another table. These are all good things, so a migration
was also added to ensure these indexes get created in
databases created by previous versions of Supermarket.

[1] http://edgeguides.rubyonrails.org/4_2_release_notes.html#active-record-notable-changes

Signed-off-by: Robb Kidd <rkidd@chef.io>