Allow omnibus test-kitchen tests to run in parallel
Previously, the exact file path to a package file had to be set
for a version. Sadly, for RHEL, that file path includes the platform
version (e.g. el6 vs el7). This updates the kitchen config and test
recipe to only require a version number for the package to be tested
and computes the file path.

package_path's default is based on the kitchen config's
synced_folders.

Note, you'll still need to have a build for a particular platform
stowed in omnibus/pkg prior to running kitchen for the package to
be available within the TK machine.

Signed-off-by: Robb Kidd <rkidd@chef.io>