Check gem versions for known vulnerabilities during security phase.
This is run during the Security phase of the Build stage because the
team concluded that in general, the findings of gems that need to be
updated is not a failure directly attributable the changes in a
changeset. As such, the change to update gem versions should come in
to the pipeline as a new change.

Signed-off-by: Robb Kidd <rkidd@chef.io>