add a metric runner service class
Provides a utility class to schedule running or rerunning fieri quality
metrics on a specific version of a cookbook, the latest version of a
cookbook, or on the latest versions of all cookbooks.

The search for all latest versions is slightly complicated so that all
the cookbook version ids can be returned in a single longish query
instead of a bombardment N+1 in Ruby/ActiveRecord to ask each cookbook
for its latest version.

Signed-off-by: Robb Kidd <rkidd@chef.io>