prevent creating metric results without a type
In testing, discovered that it was possible to create a result without
an associated metric type. This lead to views bombing on a nil being
present in an association.

This change shores up creating QualityMetric types in seeds and--as a
last resort--during the utility lookup methods for the types on the
class. Also added a not-null constraint to the schema to avoid creating
a result with no associated metric type.

Opted for .where().first_or_create!() to avoid repeating the name string
and possible future mismatched typos in the seeds.

Signed-off-by: Robb Kidd <rkidd@chef.io>