display error when runs for non-existent cookbooks are requested
Changed the return of the service class methods to be success/failure
tuples for consumers of service to deal with errors without exception
handling. (i.e. bad user input is not exceptional)

Added guards for when metric runs are requested and fieri is disabled.

Added errors for missing cookbooks or cookbook versions for the methods
that take potentially mistaken user input.

Internationalize, yo!

Signed-off-by: Robb Kidd <rkidd@chef.io>