make the API item limit configurable
Use API_ITEM_LIMIT environment variable to set the upper limit on the
number of items to return for API requests to a Supermarket instance.

Defaults to the previously hard-coded value of 100 if the env var is not
present. Also, the cookbook internal to the omnibus package has the same
default set in the section of the attributes that get set as env vars.

Signed-off-by: Robb Kidd <rkidd@chef.io>