make sidekiq optionally connect to an alternate redis
Adds an initializer for Sidekiq to configure which Redis instance to use
for job queuing. Defaults to a common one REDIS_URL that will be shared
with Rails caching and Rollout feature flag checking. Shared is less
than ideal in high traffic production environments, so this adds the
option to set a different Redis URL for Sidekiq.

https://github.com/mperham/sidekiq/wiki/Using-Redis#using-an-initializer

Signed-off-by: Robb Kidd <rkidd@chef.io>