provide a fallback for when FQDN is indeterminate
Occasionally ohai fails to figure out a host's FQDN (e.g. hostname -f
fails). Calling .downcase on nil causes much woe and gnashing of teeth.
Therefore, we'll fallback to a host's hostname for the FQDN, much like
the workaround used in many wrapper cookbooks of making an entry in
/etc/hosts for the hostname on the host's IP address if none is present.

Production Supermarkets are expected to provide an override for this
value that matches a supplied SSL certificate. Defaults are mostly for
development and test environments.

Signed-off-by: Robb Kidd <rkidd@chef.io>