prevent an exception if the user's public key is empty
If something has gone horribly awry with OC-ID and the oauth tango has
not given Supermarket a public key to associate with the user, this
stops the exception when attempting the produce a fingerprint for the
key that's really nil.

If the key is missing, say so on the user's profile page with something
to try to fix it.

Signed-off-by: Robb Kidd <rkidd@chef.io>