fix collaborator groups
With the flattening of the incoming user_ids, discovered a bug in adding
collaborators through groups. Users getting added via a group should be
added as collaborators even if they are already a direct collaborator on
a resource.

Added new direct_collaborator[s | _users] to Cookbook and Tool to
constrain list to only those users who have been added directly as
collaborators (not through a group). Then used these direct lists as the
filter for ineligible ids.

Signed-off-by: Robb Kidd <rkidd@chef.io>