fix fieri authentication
Use strong parameters to ensure a fieri_key is present in job requests
and matches the key.

This change includes a bit of a refactoring on the incoming parameters.
Instead of three individual `cookbook_` param fields, they have been
moved to live under a `cookbook` key. This makes checking for their
presence a little easier and passing the resulting hash much clearer.

Signed-off-by: Robb Kidd <rkidd@chef.io>