Fix the activity feed, give credit where it is due
As described in Issue #1509 the fix for Issue #965 (PR #1293)
shows the cookbook owner's avatar while displaying the name
of the person publishing the version.

This commit updates the message as per discussion on #1509:

> 👩 Chef: rsyslog v5.1.0 released by 👱 Timothy Smith

It also adjusts the feed to read the same way, though
without the leading leading avatar and without the trailing
publisher:

> Chef: rsyslog v5.1.0 released

but leaves the publisher as the author of the the atom
entry.  This strikes a balance between owner/maintainer
and publisher, so that the owner/maintainer's name still
leads the atom entry's subject, but the atom entry is
authored/published by the same person who published
the cookbook version.

Signed-off-by: David Alpert <david@spinthemoose.com>