embed the fieri engine gem into supermarket
This is to admit that the fieri engine "service" is really an engine
that runs within the running supermarket app itself. Until we want to
run it separately, it will be better to place the gem source somewhere
under the supermarket application source to make possible packaging the
app as a single thing without path shenanigans.

Signed-off-by: Robb Kidd <rkidd@chef.io>