stop interpolating in the default attributes
This is the start of removing interpolated values from the omnibus'
cookbook attribute defaults. Interpolating the default chef_server_url
into chef_identity_url in the defaults caused no end of trouble with
onprem installations. The application URL helpers already perform
run-time interpolation with the chef_server_url when no env var is
present. So we'll leave these attributes unset which will omit them from
the list of things that get exported as env vars.

Administrators still have the option to set a value for these attributes
which will set the associated env var and will override in the Rails
app.

Signed-off-by: Robb Kidd <rkidd@chef.io>