validate a cookbook version against chef's required format
A version for a cookbook must be either X.Y.Z or X.Y.[1] Up until this
change, the validation for a CookbookVersion's version was against
SemVer rules. SemVer allows for release labels or build info to appear
after the version number itself. Chef does not!

[1] https://docs.chef.io/cookbook_versions.html#constraints

Here, we add a new ChefVersionValidator that reuses the existing version
validation logic of the Chef gem in ::Chef::Version. Supermarket is
already relying on the Chef gem to validate version constraints set for
cookbook dependencies and supported platform.

Fixes #1714 to return an error to any client if a cookbook is uploaded
with an invalid version.

Because cookbook versions are _Chef_ and not _Semantic_ versions, remove
the Semverse dependency altogether and use Chef::Version instead.

Signed-off-by: Robb Kidd <rkidd@chef.io>