move the building of a dev omnibus package to a new cookbook
TL;DR: ensure that the omnibus-internal cookbook for
`reconfigure` does not depend on gems, directly or through a dependency
cookbook.

Fixes #1712

This moves the logic and dependencies of building an omnibus package in
a development environment to a new cookbook. This separates building
from the `omnibus-supermarket` cookbook internal to the omnibus
package and what drives `supermarket-ctl reconfigure`.

Up to this commit, `omnibus-supermarket` depended on the `omnibus`
cookbook. Its `build_package` recipe required the `omnibus` cookbook to
setup a Test Kitchen VM with the kit necessary to build.

The `omnibus` cookbook requires the `pkg-config` gem. In Nov 2017, the
installation of this gem moved from a converge-time `chef_gem` resource
to a very-early-in-the-chef-run-all-the-time gem metadata entry. Gems
declared in metadata get `bundle install`ed at the beginning of a chef
run regardless of whether the recipes/resources that use them are
called.

This early-in-the-chef-run `bundle install` was causing `supermarket-ctl
reconfigure` to fail in air-gapped environments where the host is unable
to reach rubygems.org. The omnibus' chef-client run does not use a
system's chef configuration (e.g. `/etc/chef/client.rb`) and it is
prohibitively annoying to reach into the omnibus-generated `solo.rb`
chef configuration to point it at an alternative gem server.

Signed-off-by: Robb Kidd <rkidd@chef.io>