convert omnibus serverspec to inspec
Initially the conversion to Inspec was to add a check that the cookbooks
that drive ctl-reconfigure do not depend on gems. In the end that check
was implemented in the omnibus build itself.

But the conversion to Inspec is still valuable and can be built upon. So
let's keep it!

Signed-off-by: Robb Kidd <rkidd@chef.io>