disable remote fetching in Paperclip
Upgrade to Paperclip 5.2+. There is a CVE[1] in Paperclip v5.1.0.[1]

[1] https://nvd.nist.gov/vuln/detail/CVE-2017-0889

***** Supermarket was not vulnerable to this CVE. *****

The `tarball` parameter given for an upload is validated by
CookbookUpload::Parameters and CookbookUpload::Archive before a hand-off
to ActiveRecord+Paperclip for processing as an attachment.

This update to the Paperclip initializer is intended to prevent future
accidental support for remote fetching if the logic of CookbookUpload
changes at some point and allows a URL string through to the persistence
layer. Paperclip 5.2.0 disabled these remote adapters. Leaving the
initializer disabling in case that behavior changes.

Added tests to catch regressions.

Signed-off-by: Robb Kidd <rkidd@chef.io>