fail build if omnibus-internal cookbooks depend on gems
Calls out to rubygems.org will fail in airgapped environments and it is
prohibitively annoying for administrators to configure an omnibus'
ctl-reconfigure chef-client config to use an internal gem source.

Adding a gem dependency check for any cookbooks used by the cookbook
that drives reconfigure and fail the omnibus build if present.

Signed-off-by: Robb Kidd <rkidd@chef.io>