appease Rubocop: Rails/InverseOf
https://rubocop.readthedocs.io/en/latest/cops_rails/#railsinverseof

Provided `inverse_of:` for associations that had options that made it
difficult/impossible for ActiveRecord to figure out the other direction.

Cookbook.collaborators was the most difficult because it took plumbing
the naming philosophy of the model. Collaborator is a link table/model
connecting a thing that can be collaborated on with the people who can
collaborate on it. A better name for the model might be "Collaboration"
much like a link between people and groups could be called a
"Membership". Formally renaming this class is way beyond the scope of
a Rubocop appeasement, so we've just added some documentation bread
crumbs for the people of the future.

Cookbook.replaces is a new association to return the collection of
cookbooks that have deprecated in favor of a given cookbook. This
represents the inverse of Cookbook.replacement.

Co-authored-by: Jennifer Davis <iennae@gmail.com>
Signed-off-by: Robb Kidd <rkidd@chef.io>