appease Rubocop: Style/SafeNavigation
> Use safe navigation (&.) instead of checking if an object exists
> before calling the method.
>    url.match(%r{(?<=github.com\/)[\w-]+\/[\w-]+}).to_s unless url.nil?
>    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

True, yes, we could use the new, inscrutable `try` method navigation.

Instead, opted for making the method that performs the source_repo_url
lookup always return a string.

Added a spec for the SourceRepoWorker class that is the parent of the
other workers that care about a cookbook's source repository.

Signed-off-by: Robb Kidd <rkidd@chef.io>