catch new ArgumentError exception for invalid tarballs
Ruby 2.5.1 sync'd up with RubyGems 2.7.6. RubyGems 2.7.6 added a strict
octal check[1] for entries in tarballs which occurs earlier than the
previous exception thrown for Gem::Package:TarInvalidError. This adds
the new exception type thrown to our rescue to add our error message
text to what is displayed to the user.

[1] https://github.com/rubygems/rubygems/commit/92e98bf8f810bd812f919120d4832df51bc25d83

Signed-off-by: Robb Kidd <rkidd@chef.io>