switch to using Redcarpet's Safe HTML renderer
Fixes a bug in rendering links and images with unsafe protocols in the
display of user-provided markdown.

Redcarpet::Render::Safe[^1] is the class provided by the library for
rendering markdown provided by users (so not trustworthy). As of
redcarpet v3.4.0, Safe's defaults are

* escape HTML characters
* skip rendering links and images with unsafe URLs
* not to include a code block's language in the tag

So we've switched to the Safe renderer and expanded the specs testing
for preventing XSS injections into the rendered HTML.

Thanks to Sundarr Insa for reporting the behavior!

[^1]: https://github.com/vmg/redcarpet/blob/v3.4.0/lib/redcarpet.rb#L25-L60

Signed-off-by: Robb Kidd <rkidd@chef.io>