MSYS-878 set session cookie to secure when sent over an HTTPS connection
All omnibus installs run with a Rails.env of "production" but not all
installs run with SSL enabled. So we'll need to determine SSLiness from
the URL scheme/protocol the system is configured to use for linking back
to itself.

Placed the logic in a library class method to (1) give the test a
meaningful name and (2) make it possible to test various return values
outside the initializer.

Co-authored-by: Robb Kidd <rkidd@chef.io>
Co-authored-by: Vasu1105 <vasundhara.jagdale@msystechnologies.com>
Signed-off-by: Robb Kidd <rkidd@chef.io>
Signed-off-by: Vasu1105 <vasundhara.jagdale@msystechnologies.com>