replace periodic job scheduler
sidetiq[1] was heretofore our periodic background job scheduler. It was
deprecated a couple years ago, but still worked for Supermarket's
purposes.

sidetiq required the use of celluloid which has also been deprecated.
The version of celluloid require was old and required the use of a
celluloid subcomponent that was brought in via git submodules.

There is much sadness in everything above this line.

sidekiq-cron[2] has been a go-to replacement for sidetiq in other projects
since its maintainer declared it unmaintained. Here, we replace sidetiq
with sidekiq-cron. This required:

* updated the Gemfile
* removing the require shenanigans for celluloid in application.rb
* setting the schedule for Supermarket's recurring jobs in the Sidekiq
  startup config in the initializer; putting in Sidekiq's startup
  prevents it from being called and attempting a write to Redis whenever
  the Rails app is initialized (e.g. precompiling assets, running tests,
  opening the console).
* removing the Sidetiq pieces from the scheduled jobs

The Jobs:

SitemapRefreshWorker was simple; remove Sidetiq things.

OauthTokenRefreshScheduleWorker was more invasive. sidekiq-cron does not
have the same schedule backfill feature, so there is no last_occurence
or current_occurence to work with. This job now assumes that it is run
every five minutes to refresh any tokens. The querying of account with
access tokens expiring soon moves to the Account class. This query gets
to define what "soon" means. The job itself becomes much simpler by
relying on this query and scheduling the token refresh for any account
returned.

[1] https://github.com/endofunky/sidetiq
[2] https://github.com/ondrejbartas/sidekiq-cron

Co-Authored-by: Paul Welch <pwelch@chef.io>
Signed-off-by: Robb Kidd <rkidd@chef.io>