expand on a Habitat plan
Unicorn start up gets a bit of a tune-up. We'll put "rails server" into
the Procfile as the "web" command. The unicorn-rails gem handles hooking
unicorn in as the web server to start up and also handles loading a
config file based on the name of the Rails environment.

And then Joshua told me to stop writing ...

Signed-off-by: Robb Kidd <rkidd@chef.io>

Update habitat packaging for latest development

This commit removes the toplevel habitat directory, as we'll create
the package from the `src/supermarket` directory, so that the
scaffolding-ruby magic works correctly. We also need to create /root
in our studio, so we put that in .studiorc (see
habitat-sh/habitat#5569 for background on that). Finally, we need to
add the tzinfo-data gem to the Gemfile to get the package to build
correctly.

Signed-off-by: Joshua Timberman <joshua@chef.io>