upgrade FactoryGirl to FactoryBot
Because the latest is a good thing to have.

Because it's the right thing to do.

Because there will probably be some factory shenanigans with a Rails 5.0
to 5.1 upgrade.

Also moves factorybot and faker to the development & test group because
they are sometimes handy in the development environment for cobbling up
test data.

Also also alphabetized the test group because rubocop likes that.

Also also also, resolve deprecation warning about static attributes in
factories.

Signed-off-by: Robb Kidd <rkidd@chef.io>