upgrade NodeJS in the omnibus install
Otherwise the default version is very old. NodeJS is not used in any
runtime, user-facing service of Supermarket. It is a build dependency
for precompiling the Javascript assets in the web application. It
remains in the install for private Supermarkets whose administrators
wish to customize the branding which requires recompiling the assets.

Signed-off-by: Robb Kidd <rkidd@chef.io>