implement RFC072: only allow admins to remove cookbooks and cookbook versions
Chef RFC072[1] was accepted as a plan for Supermarket to prevent cookbook
artifacts from disappearing. These changes are a beginning: prevent a
cookbook owner from removing any version of a cookbook or the entire
cookbook itself. This change does not address the proposed hiding
behavior in the RFC.

For backward compatibility, a new configuration option has been added to
enable the previous behavior to allow artifact removal by owners. The
option is set via an environment variable managed by the omnibus
install. The current default is to enable it with the intention to
change the default to disabled in a future major release.

[1] https://github.com/chef/chef-rfc/blob/f8250a4746d2df530b605ecfaa2dc5ae9b7dc7ff/rfc072-artifact-yanking.md

Co-authored-by: Jon Morrow <jmorrow@chef.io>
Co-authored-by: Marc A. Paradise <marc.paradise@gmail.com>
Co-authored-by: tyler-ball <tball@chef.io>
Signed-off-by: Robb Kidd <rkidd@chef.io>