require POSTs during sign-in (CVE-2015-9284)
CVE-2015-9284 was a bit of a doozy. Essentially, a service provider
(oauth "client") ought to accept authentication initiations by POST with
a token from itself to protect against cross-site forgeries. Details so
far on omniauth/omniauth issue 809.

This change does the following:

1. upgrades omniauth and adds a new omniauth-rails_csrf_protection gem

The new gem twiddles with OmniAuth's middleware to require POSTing to
the auth endpoints it adds to the app and wires up checking for a valid
token during the middleware processing. The omniauth gem itself is
framework agnostic, so the current solution is a new Rails-flavored
library.

2. updates all Supermarket's links to OAuth2 providers to use POST

Pretty easy update to link_to "buttons" in the views.

3. display a page on /sign-in with a login button

Because the endpoint that actually signs a user is /auth/chef-oauth2, a
route and sessions controller action was made to redirect there for a
friendlier URL path. But you can't redirect POSTs, yo. So, now that
we're POSTing, if someone lands on /sign-in some how, display a button
for logging in via Chef Server. This can be the place where more buttons
are added if we ever decide to implement multi-auth.

4. redirects to /sign-in changed from 301 to 302

Browsers take permanent redirects at their word and seem to never forget
them.  Nothing is really permanent on the internet.

Signed-off-by: Robb Kidd <rkidd@chef.io>