Add s3_private_objects info to omnibus attributes
As more people roll out restrictions on public ACLs for their S3 buckets
to avoid being the next headline we're likely to see people running into
problems with their internal supermarket.  The info they need is here in
the repository but it's not easy to find. This change is to surface that
in a place people can find directly from the cookbook README.

Obvious fix to add documented setting to default config reference.[1]

[1] https://github.com/chef/supermarket/pull/1819#issue-303037189

Co-authored-by: Mark Bainter <mbainter+github@gmail.com>
Signed-off-by: Robb Kidd <rkidd@chef.io>