install inspec as a gem for the -ctl command
The inspec software definition in omnibus-software does not (yet?)
also install the inspec-bin gem to get binstub executables created.
Because the supermarket-ctl command is the thing that needs inspec as a
command to shell out to _and_ it is managed as a software definition
_and_ the software definition is already bundle-installing, I've added
inspec and inspec-bin to the Gemfile used by the software definition.

This results in the latest _release_ inspec, as opposed to whatever is
current on the master branch, getting installed within Supermarket as an
executable. Licensing is handled at the call site in the -ctl command
that uses inspec.

Since we're bring in the latest, we need to accept the license for the
duration of the -ctl test.

Signed-off-by: Robb Kidd <rkidd@chef.io>