override reconfigure command to accept chef license
By adding a `reconfigure` command to the `general` category, we can
override the one provided by omnibus-ctl that performs license checking
for a project. It was decided that the use of chef-client within
Supermarket's -ctl commands was a "library use," so a non-persistent
acceptance at the call sites is acceptable.

This reconfigure command looks much like omnibus-ctl's, but with the
license review and acceptance removed. This command still uses the
`run_chef` method and passes in the non-persistent acceptance as an
argument.

This change also updates omnibus-ctl because it was woefully behind.

Signed-off-by: Robb Kidd <rkidd@chef.io>