generate secrets.json with warning messages
As described in #1839, secrets.json now wins for config values.

If a secret_key_base has been provided by an existing secrets.json file,
no problem so long as that file is the same across multiple web hosts.

If no secrets.json file is present and a secret_key_base value was
provided via supermarket.rb or supermarket.json, a warning is
presented that this value should really be in secrets.json and
the secrets.json file is generated with that value and used from there.

If no secrets.json file is present and no secret_key_base value was
provided, a warning is presented that a value is being generated. If
the installation has only one web host, the generated value is fine. If
the installation has 2 or more web hosts, the administrators are
required to ensure that the secrets.json file matches across those
hosts.

The omnibus installer is not empowered to resolve a discrepancy in secrets
values across multiple hosts. That is a secrets management/orchestration
problem outside the purview of omnibus.

Signed-off-by: Robb Kidd <rkidd@chef.io>