enable Strict-Transport-Security within web app itself
nginx was already redirecting incoming connections from HTTP to HTTPS if
nginx->force_ssl was set to true. But the web app was not including the
STS header. This change adds a new environment variable FORCE_SSL with a
value equal to the one set for nginx. If true, Rails' config.force_ssl
is also set to true which will cause responses to include the header.

Updated the install-check InSpec test to confirm the presence of the
HTTP Strict-Transport-Security header when force_ssl is set for the
installation.

Signed-off-by: Robb Kidd <rkidd@chef.io>