fix a file reader left at the end of the file contents
Having read the contents of the tarball file, it needed to be rewound so
that using it again in the test POST didn't result in an empty file being
processed in the test environment. This appeared to only be a problem in
the tests.

Not 100% positive of the change in Rails or the test framework that this
behavior changed, but I suspect the same file object instance was being
passed from the spec and into the application during testing.

Signed-off-by: Robb Kidd <rkidd@chef.io>