update spec to allow running by itself
This spec uses a helper method. When run with the entire suite, something
else (probably the vcr_helper) was loading the helper. When run by
itself or only along with the feature specs, this would fail because the
helper method wasn't defined.

So, include the helper that defines that method so the spec can run on
its lonesome. This is only a little dirty.

Signed-off-by: Robb Kidd <rkidd@chef.io>