annotate existing migrations with legacy Rails version number
Rails 5.0 introduced a Migration API.[1] ActiveRecord::Migration classes
are Rails version specific.

Rails 5.0 was pretty forgiving about migration classes that didn't have
a version annotation. Rails 5.1 is not. These changes are in preparation for
new migrations to be added and so only touch existing files.

Incidentally, it is `rake db:migrate` that reeeeally wants these old
migration classes to declare for which Rails version they were written.
`rails db:migrate` seemed to accept them just fine. But we rake
everywhere, so let's add the annotations to keep up with the framework.

[1] https://github.com/rails/rails/pull/21538

Signed-off-by: Robb Kidd <rkidd@chef.io>