new CtlCommandHelper method for rails commands
`console` and `make-admin` commands were failing on some platforms as a
result of HOME not being set to a directory to which the command could
write. Both need the same preamble setup to get the stars in the correct
alignment for Rails commands to succeed, so the preamble moved to the
existing helper class. Add a HOME entry to that preamble.

`quality-metrics` needed the same treatment.

Signed-off-by: Robb Kidd <rkidd@chef.io>