turn s3 config audit into s3 config generator
In an effort to avoid testing the behavior of Paperclip as a library
with a matrix of integration tests, opted to test that we are generating
a "correct" Paperclip configuration and trust that the library does what
it says it does.

The paperclip initializer had gone bananas with branching a long while
ago and initializers are wildly difficult to test variations on. This
change turns the S3 config auditing class into an S3 config generator
to combine audit with transforming environment variables into a
Paperclip config. The majority of the branching logic moves to a
to_paperclip_options method where the branches and output can be tested
with specs.

Once tests were in place on the previous output, the body of the method
got an overhaul to replace merging hashes with simpler key assignments.

Signed-off-by: Robb Kidd <rkidd@chef.io>