disambiguate columns used in query ordering
Queries using the named ordering options and also joins with other
tables were failing with:

  ActionView::Template::Error: PG::AmbiguousColumn:
  ERROR:  column reference "id" is ambiguous
  LINE 1: SELECT DISTINCT cookbook_followers_count, id AS alias_0, "c…
                                                    ^

Database query was traced to these ordering options in the Cookbook
scopes.

Added tests for the API route and query parameters that triggered these
errors. Added tablenames to the column references to pass the tests.

Signed-off-by: Robb Kidd <rkidd@chef.io>