fix error page routing and display
Rails 5.1 deprecated using `env` in controllers in favor of using
`request.env`. This change switches to that and removes some methods
that are not currently used anywhere.

The result of the `env` calls being in place is that the exception show
page rendering would itself cause an exception. Thus, all errors become
500 errors which heightens the suspicion of upstream proxies and load
balancers. The application logs also get filled with stacktraces for
simple non-matched route errors which no body has time for.

Signed-off-by: Robb Kidd <rkidd@chef.io>